Move template related views into templates app
